requirements.txt
# Core Django and DRF
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication & Authorization
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Fix for pkg_resources compatibility issue
setuptools==69.5.1

# Image Processing
Pillow==10.0.1

# Environment Variables
python-decouple==3.8

# Documentation
drf-spectacular==0.26.5

# Filtering and Pagination
django-filter==23.3

# Caching
redis==5.0.1
django-redis==5.4.0

# File Storage (for production)
django-storages==1.14.2
boto3==1.29.7

# Development Tools
django-debug-toolbar==4.2.0

# Validation
django-phonenumber-field==7.2.0
phonenumbers==8.13.25

# Utilities
django-extensions==3.2.3
celery==5.3.4
